BackWallTransparent
The back wall is now spawned in by each player, a solid textured wall by the chef, and a transparent wall by the rat. Not an elegant solution, but post processing does not work for transparency. Also added a copy of the back wall with no owner, set to only owner see, so the back wall can be seen in the editor, but is not rendered at run time.

Any other walls or objects with a similar effect: in each character blueprint, when the character is possessed, spawn in a blueprint of the object at the desired transform