Improved outlines + slight map change
Switched grey and blue table.
Minigames now have material and outlines applied, much better.

Also disabled the audio volumes so no music fading in and out

also also slightly stopped the audio clipping on the fridge. however it still sounds like clipping. this is because the sound actually plays a bunch of times at once. derek needs to fix this